[BSE-883] Support ReserveTable with bitmap (#5929)
* Add ReserveTable with bitmap

* Add ReserveTable with bitmap [run CI]

* Address comments

* Fix [run CI]

* [run CI]

* [run CI]

* [run CI]

* Remove u_int64_t [run CI]

* [run CI]

+++ bodo/libs/_chunked_table_builder.cpp
@@ -371,8 +371,8 @@ void ChunkedTableBuilder::AppendBatch(
     const std::shared_ptr<table_info>& in_table,
     const std::vector<bool>& append_rows) {
     // Convert bool vector into indices vector
-    u_int64_t num_append_rows =
-        std::accumulate(append_rows.begin(), append_rows.end(), (u_int64_t)0);
+    size_t num_append_rows =
+        std::accumulate(append_rows.begin(), append_rows.end(), (size_t)0);
     std::vector<int64_t> idxs;
     idxs.reserve(num_append_rows);
